prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
   */
 int main(int fib)
 {
-	int  fib1 = 1, fib2 = 1, swap; 
+	unsigned long int  fib1 = 1, fib2 = 1, swap; 
 
 
 	for (fib = 1; fib <= 98; fib++)
